+++ common/facts_utils.py
